Generate dependency types of relations
Problem
===

Previously it has lack of generated classes. Because it doesn't make
sure that relation's type exists in RBS level.
For example, it generates `def user: () -> User` from `belongs_to :user`,
but it does not check the existence of `User` class.

I encountered this problem with Active Hash. In the example, if `User`
is a subclass of ActiveHash::Base, the `User` class is not generated by
RBS Rails so it crashes.

Solution
===

Generate a polyfill for a non existing class. It fixes the problem.